finish

+++ App/delete/hapus_data.php
@@ -1,25 +1,21 @@
 <?php 
 include('../../conf/config.php');
-// Mengakses data dari $_GET dengan pengecekan agar tidak terjadi 'undefined array key'
-$id = $_GET['id'];
 
+if($_GET['kode'] == 'karyawan'){
+    $id = $_GET['id'];
+    $query = mysqli_query($koneksi,"DELETE FROM tb_karyawan WHERE id='$id' ");
+    header('Location: ../index.php?page=data-karyawan');
+}
 
-//data admin
-$id_admin = $_GET['id_admin'];
+if($_GET['kode'] == 'admin'){
+    $id_admin = $_GET['id_admin'];;
+    $query = mysqli_query($koneksi,"DELETE FROM tb_admin WHERE id_admin='$id_admin' ");
+    header('Location: ../index.php?page=data-admin');
+}
 
-//data member
-$id = $_GET['id'];
-
-
-// Menampilkan data dengan aman
-//echo htmlspecialchars($nama) . "<br>";
-//echo htmlspecialchars($email) . "<br>";
-//echo htmlspecialchars($jabatan) . "<br>";
-//echo htmlspecialchars($status) . "<br>";
-$query = mysqli_query($koneksi,"DELETE FROM tb_karyawan WHERE id='$id' ");
-header('Location: ../index.php?page=data-karyawan');
-$query = mysqli_query($koneksi,"DELETE FROM tb_admin WHERE id='$id_admin'");
-header('Location: ../index.php?page=data-admin');
-$query = mysqli_query($koneksi,"DELETE FROM tb_member WHERE id='$id'");
-header('Location: ../index.php?page=data-admin');
+if($_GET['kode'] == 'member'){
+    $id = $_GET['id'];
+    $query = mysqli_query($koneksi,"DELETE FROM tb_member WHERE id='$id' ");
+    header('Location: ../index.php?page=data-member');
+}
 ?>
